+++ service/src/main/java/org/ehrbase/dao/access/jooq/ContextAccess.java
@@ -41,7 +41,8 @@ import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.dao.access.interfaces.I_CompositionAccess;
 import org.ehrbase.dao.access.interfaces.I_ContextAccess;
 import org.ehrbase.dao.access.interfaces.I_DomainAccess;
-import org.ehrbase.dao.access.interfaces.I_PartyIdentifiedAccess;
+import org.ehrbase.dao.access.jooq.party.PersistedObjectId;
+import org.ehrbase.dao.access.jooq.party.PersistedPartyProxy;
 import org.ehrbase.dao.access.support.DataAccess;
 import org.ehrbase.jooq.pg.tables.records.*;
 import org.ehrbase.serialisation.RawJson;
@@ -53,6 +54,10 @@ import org.jooq.exception.DataAccessException;
 import java.sql.PreparedStatement;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.temporal.ChronoField;
+import java.time.temporal.TemporalAccessor;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -164,7 +169,7 @@ public class ContextAccess extends DataAccess implements I_ContextAccess {
                         .and(PARTICIPATION_HISTORY.SYS_TRANSACTION.eq(transactionTime)))
                 .forEach(record -> {
                     //retrieve performer
-                    PartyProxy performer = I_PartyIdentifiedAccess.retrievePartyIdentified(domainAccess, record.getPerformer());
+                    PartyProxy performer = new PersistedPartyProxy(domainAccess).retrieve(record.getPerformer());
 
 
                     DvInterval<DvDateTime> startTime = convertDvIntervalDvDateTimeFromRecord(eventContextHistoryRecord);
@@ -203,12 +208,9 @@ public class ContextAccess extends DataAccess implements I_ContextAccess {
 
     private static PartyIdentified getPartyIdentifiedFromRecord(PartyIdentifiedRecord partyIdentifiedRecord, List<DvIdentifier> identifiers) {
         PartyIdentified healthCareFacility;
-        PartyRef partyRef;
+        PartyRef partyRef = null;
         if (partyIdentifiedRecord.getPartyRefValue() != null && partyIdentifiedRecord.getPartyRefScheme() != null) {
-            GenericId genericID = new GenericId(partyIdentifiedRecord.getPartyRefValue(), partyIdentifiedRecord.getPartyRefScheme());
-            partyRef = new PartyRef(genericID, partyIdentifiedRecord.getPartyRefNamespace(), partyIdentifiedRecord.getPartyRefType());
-        } else {
-            ObjectId objectID = new HierObjectId("ref");
+            ObjectId objectID = new PersistedObjectId().fromDB(partyIdentifiedRecord);
             partyRef = new PartyRef(objectID, partyIdentifiedRecord.getPartyRefNamespace(), partyIdentifiedRecord.getPartyRefType());
         }
         healthCareFacility = new PartyIdentified(partyRef, partyIdentifiedRecord.getName(), identifiers.isEmpty() ? null : identifiers);
@@ -231,7 +233,7 @@ public class ContextAccess extends DataAccess implements I_ContextAccess {
 
         //Health care facility
         if (eventContext.getHealthCareFacility() != null) {
-            UUID healthcareFacilityId = I_PartyIdentifiedAccess.getOrCreateParty(this, eventContext.getHealthCareFacility());
+            UUID healthcareFacilityId = new PersistedPartyProxy(this).getOrCreate(eventContext.getHealthCareFacility());
 
             eventContextRecord.setFacility(healthcareFacilityId);
         }
@@ -284,7 +286,7 @@ public class ContextAccess extends DataAccess implements I_ContextAccess {
                 }
 
                 performer = (PartyIdentified) setPerformer;
-                UUID performerUuid = I_PartyIdentifiedAccess.getOrCreateParty(this, performer);
+                UUID performerUuid = new PersistedPartyProxy(this).getOrCreate(performer);
                 //set the performer
                 participationRecord.setPerformer(performerUuid);
                 participations.add(participationRecord);
@@ -492,7 +494,7 @@ public class ContextAccess extends DataAccess implements I_ContextAccess {
         //get the participations
         getContext().fetch(PARTICIPATION, PARTICIPATION.EVENT_CONTEXT.eq(eventContextRecord.getId())).forEach(record -> {
             //retrieve performer
-            PartyProxy performer = I_PartyIdentifiedAccess.retrievePartyIdentified(this, record.getPerformer());
+            PartyProxy performer = new PersistedPartyProxy(this).retrieve(record.getPerformer());
 
             DvInterval<DvDateTime> dvInterval = convertDvIntervalDvDateTimeFromRecord(record);